commit classes Debito, Credito, Pix, e interface PagAvista

+++ Loja_produtos/src/Perifericos.java
@@ -1,5 +1,5 @@
 public class Perifericos extends Produto {
-    public Perifericos(int preco){
+    public Perifericos(Double preco){
         super(preco);
     }
